♻️ [Logs] remove stubs and move the API in logs.entry
Tests will be moved in a separate commit to ease the review
